refacted mobileIndex to a separate file
