+++ dldashboard/web/src/MainPane.js
@@ -20,8 +20,6 @@ class MainPane extends React.Component {
 
   render() {
     const stateManager = this.props.stateManager;
-    console.log("main pane state manager");
-    console.log(stateManager);
 
     return (
       <div>
